Moved offline storage management functionality into new view controllerr
