+++ ownCloud/Resources/en.lproj/Localizable.strings
@@ -279,3 +279,4 @@
 "Offline files use" = "Offline files use";
 "Compacting of '%@' failed" = "Compacting of '%@' failed";
 "Delete Offline Copies" = "Delete Offline Copies";
+"Manage" = "Manage";
